compilando notes, ainda nao esta pronto, defines added

+++ addresspageview.h
@@ -22,7 +22,7 @@ class AddressPageView : public QFrame
 public:
   explicit AddressPageView(QWidget* parent = nullptr);
   QVector<Address> getAddresses() const;
-  QVector<int> getRemovedAddresses() const;
+  QVector<qlonglong> getRemovedAddresses() const;
   void setAddresses(const QVector<Address>& vAddress);
 
 public slots:
@@ -44,7 +44,7 @@ private slots:
 
 private:
   QListWidgetItem* m_currentItem;
-  QVector<int> m_vRemovedAddressId;
+  QVector<qlonglong> m_vRemovedAddressId;
   Address getAddress();
   JLineEdit* m_edCep;
   QPushButton* m_btnCep;
