+++ phonepageview.h
@@ -21,7 +21,7 @@ class PhonePageView : public QFrame
 public:
   explicit PhonePageView(QWidget* parent = nullptr);
   QVector<Phone> getPhones() const;
-  QVector<int> getRemovedPhones() const;
+  QVector<qlonglong> getRemovedPhones() const;
   void setPhones(const QVector<Phone>& vPhone);
 
 public slots:
@@ -43,7 +43,7 @@ private slots:
 private:
   Phone getPhone();
   QListWidgetItem* m_currentItem;
-  QVector<int> m_vRemovedPhoneId;
+  QVector<qlonglong> m_vRemovedPhoneId;
   QSpinBox* m_spnCountryCode;
   QSpinBox* m_spnCode;
   JLineEdit* m_edNumber;
